feat: woobler button now appears at the bottom
